requirements.txt
# Core ML stack
tensorflow-cpu>=2.13.0
keras>=2.13.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Scheduling, HTTP, utilities
APScheduler>=3.10.1
requests>=2.31.0

# Optional: enable progress bars and nice logs (safe to remove)
tqdm>=4.66.0
rich>=13.5.0

